Refactor: structure refactor for stackinbox.util_*
- Move stackinabox.utils_* into a dedicated module stackinabox.util
- Moved the existing stackinabox.utils to stackinabox.util.tools so that there are not two nearly identically named modules
- Rename stackinabox.utils_* to match the python module they are supporting (e.g stackinabox.util_httpretty is now stackinabox.util.httpretty)
- Updated documentation and tests

Partially Closes #36
