+++ docs/api/utils/httpretty.rst
@@ -5,6 +5,6 @@ HTTPretty Utility
 
 StackInABox provides support for writing tests with httpretty.
 
-.. currentmodule:: stackinabox.util_httpretty
+.. currentmodule:: stackinabox.util.httpretty
 .. autofunction:: httpretty_callback
 .. autofunction:: httpretty_registration
