+++ docs/changes/latest.rst
@@ -10,6 +10,8 @@ New
 
 Breaking Changes
 ----------------
+- Refactored the utilities for the module names to be less awkward. stackinabox.util_* is now stackinabox.util.*
+- Moved stackinabox.utils to stackinabox.util.tools
 
 Fixed
 -----
