+++ stackinabox/tests/test_responses.py
@@ -9,7 +9,7 @@ import requests
 import responses
 import six
 
-import stackinabox.util_responses
+import stackinabox.util.responses
 from stackinabox.stack import StackInABox
 from stackinabox.services.hello import HelloService
 from stackinabox.tests.utils.services import AdvancedService
@@ -24,7 +24,7 @@ def test_basic_responses():
     def run():
         StackInABox.reset_services()
         StackInABox.register_service(HelloService())
-        stackinabox.util_responses.responses_registration('localhost')
+        stackinabox.util.responses.responses_registration('localhost')
 
         res = requests.get('http://localhost/hello/')
         assert res.status_code == 200
@@ -38,7 +38,7 @@ def test_advanced_responses():
     def run():
         responses.mock.start()
         StackInABox.register_service(AdvancedService())
-        stackinabox.util_responses.responses_registration('localhost')
+        stackinabox.util.responses.responses_registration('localhost')
 
         res = requests.get('http://localhost/advanced/')
         assert res.status_code == 200
