+++ stackinabox/util/httpretty.py
@@ -1,6 +1,8 @@
 """
 Stack-In-A-Box: HTTPretty Support
 """
+from __future__ import absolute_import
+
 import logging
 import re
 
@@ -10,7 +12,7 @@ import httpretty.http
 import six
 
 from stackinabox.stack import StackInABox
-from stackinabox.utils import CaseInsensitiveDict
+from stackinabox.util.tools import CaseInsensitiveDict
 
 
 logger = logging.getLogger(__name__)
