+++ stackinabox/util/requests_mock.py
@@ -1,6 +1,8 @@
 """
 Stack-In-A-Box: HTTPretty Support
 """
+from __future__ import absolute_import
+
 import contextlib
 import functools
 import io
@@ -21,7 +23,7 @@ import requests_mock.response
 import six
 
 from stackinabox.stack import StackInABox
-from stackinabox.utils import CaseInsensitiveDict
+from stackinabox.util.tools import CaseInsensitiveDict
 
 
 logger = logging.getLogger(__name__)
