+++ stackinabox/util/tools/__init__.py
@@ -1,0 +1,1 @@
+from stackinabox.util.tools.caseinsensitivedict import CaseInsensitiveDict
