+++ stackinabox/util/tools/caseinsensitivedict.py
