--- stackinabox/utils/__init__.py
@@ -1,1 +1,0 @@
-from stackinabox.utils.caseinsensitivedict import CaseInsensitiveDict
